+++ Frontend/src/components/AlarmItem/alarmItem.animation.js
