+++ Frontend/src/components/AlarmItem/alarmItem.styles.js
